requirements.txt
flask==2.3.2
gunicorn
werkzeug

# Google Drive API dependencies
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
